Make the navbar configurable by site
Each site sets what should show in their own navbar.
The url will be the slug of the section name.
The icon can be set, otherwise it is set to the slug of the section name.
Hidden can be set on the icons, the width for the different sections
comes from what was in the original nav-top.html.

Closes #9