Improve the quality of optimized images
As I understand this script is used to optimize the uploaded images to proper dimensions. However, the quality of resized images suffer greatly. See https://beta.diagnijmegen.nl/presentations/. This should fix the issue, however I do not have the whole system running on my local, so I did not test it.